allow to use main without shell script
